[OGUI-992] Split components
